refactor: move pin function to auth.controller
