+++ controllers/auth.controller.js
@@ -14,7 +14,7 @@ const register = async (req, res, next) => {
 
         if (!nama || !email || !password || no_telp) {
             return res.status(400).json({
-                status: 400,
+                status: false,
                 message: "Bad Request"
             });
         }
@@ -23,7 +23,7 @@ const register = async (req, res, next) => {
 
         if (exist) {
             return res.status(401).json({
-                status: 401,
+                status: false,
                 message: "Email has already been used"
             });
         }
@@ -46,7 +46,7 @@ const register = async (req, res, next) => {
         delete user.password;
 
         return res.status(201).json({
-            status: 201,
+            status: true,
             message: "Created",
             data: user
         });
@@ -66,7 +66,7 @@ const login = async (req, res, next) => {
 
         if (!email || !password) {
             return res.status(400).json({
-                status: 400,
+                status: false,
                 message: "Bad Request"
             });
         }
@@ -74,7 +74,7 @@ const login = async (req, res, next) => {
         const user = await prisma.user.findFirst({ where: { email: email } });
         if (!user) {
             return res.status(401).json({
-                status: 401,
+                status: false,
                 message: "invalid email or password!"
             });
         }
@@ -82,7 +82,7 @@ const login = async (req, res, next) => {
         const isPasswordCorrect = await bcrypt.compare(password, user.password);
         if (!isPasswordCorrect) {
             return res.status(401).json({
-                status: 401,
+                status: false,
                 message: "invalid email or password!"
             });
         }
@@ -92,7 +92,7 @@ const login = async (req, res, next) => {
         const token = jwt.sign(user, process.env.JWT_SECRET);
 
         return res.status(200).json({
-            status: 200,
+            status: true,
             message: "OK",
             data: { ...user, token }
         });
@@ -114,7 +114,7 @@ const createPin = async (req, res, next) => {
 
         if (!user) {
             return res.status(401).json({
-                status: 401,
+                status: false,
                 message: "invalid email or password!"
             });
         }
@@ -132,9 +132,89 @@ const createPin = async (req, res, next) => {
         });
 
         return res.status(200).json({
-            status: 200,
+            status: true,
             message: "OK",
             data: req.user
+        });
+    } catch (error) {
+        next(error);
+    }
+};
+
+/**
+ * @param {import("express").Request} req
+ * @param {import("express").Response} res
+ * @param {import("express").NextFunction} next
+ */
+const pinValidation = async (req, res, next) => {
+    try {
+        const { pin } = req.body;
+
+        if (!pin) {
+            return res.status(400).json({
+                status: false,
+                message: "Bad Request"
+            });
+        }
+
+        const users = await prisma.user.findUnique({
+            where: {
+                id: req.user.id
+            }
+        }).Profile();
+
+        if (users.pin !== Number(pin)) {
+            return res.status(401).json({
+                status: false,
+                message: "Unauthorized"
+            });
+        }
+
+        return res.status(200).json({
+            status: true,
+            message: "OK"
+        });
+    } catch (error) {
+        next(error);
+    }
+};
+
+/**
+ * @param {import("express").Request} req
+ * @param {import("express").Response} res
+ * @param {import("express").NextFunction} next
+ */
+const forgotPin = async (req, res, next) => {
+    try {
+        const { password, pin } = req.body;
+        const users = await prisma.user.findUnique({
+            where: {
+                id: req.user.id
+            }
+        });
+
+        const passwordCorrect = bcrypt.compare(password, users.password);
+
+        if (!passwordCorrect) {
+            return res.status(401).send({
+                status: false,
+                message: "Unauthorized"
+            });
+        }
+
+        const usersPin = await prisma.profile.update({
+            data: {
+                pin: pin
+            },
+            where: {
+                id: users.id
+            }
+        });
+
+        return res.status(201).json({
+            status: true,
+            message: "Created",
+            data: usersPin.pin
         });
     } catch (error) {
         next(error);
@@ -149,7 +229,7 @@ const createPin = async (req, res, next) => {
 const whoami = async (req, res, next) => {
     try {
         return res.status(200).json({
-            status: 200,
+            status: true,
             message: "OK",
             data: req.user
         });
@@ -162,5 +242,7 @@ module.exports = {
     register,
     login,
     whoami,
-    createPin
+    createPin,
+    forgotPin,
+    pinValidation
 };
