+++ controllers/user.controller.js
@@ -18,7 +18,7 @@ const profile = async (req, res, next) => {
         delete users.password;
 
         return res.status(200).json({
-            status: 200,
+            status: true,
             message: "OK",
             data: users
         });
@@ -56,89 +56,9 @@ const updateProfile = async (req, res, next) => {
         });
 
         return res.status(200).json({
-            status: 200,
+            status: true,
             message: "OK",
             data: profile
-        });
-    } catch (error) {
-        next(error);
-    }
-};
-
-/**
- * @param {import("express").Request} req
- * @param {import("express").Response} res
- * @param {import("express").NextFunction} next
- */
-const pinValidation = async (req, res, next) => {
-    try {
-        const { pin } = req.body;
-
-        if (!pin) {
-            return res.status(400).json({
-                status: 400,
-                message: "Bad Request"
-            });
-        }
-
-        const users = await prisma.user.findUnique({
-            where: {
-                id: req.user.id
-            }
-        }).Profile();
-
-        if (users.pin !== Number(pin)) {
-            return res.status(401).json({
-                status: 401,
-                message: "Unauthorized"
-            });
-        }
-
-        return res.status(200).json({
-            status: 200,
-            message: "OK"
-        });
-    } catch (error) {
-        next(error);
-    }
-};
-
-/**
- * @param {import("express").Request} req
- * @param {import("express").Response} res
- * @param {import("express").NextFunction} next
- */
-const forgotPin = async (req, res, next) => {
-    try {
-        const { password, pin } = req.body;
-        const users = await prisma.user.findUnique({
-            where: {
-                id: req.user.id
-            }
-        });
-
-        const passwordCorrect = bcrypt.compare(password, users.password);
-
-        if (!passwordCorrect) {
-            return res.status(401).send({
-                status: 401,
-                message: "Unauthorized"
-            });
-        }
-
-        const usersPin = await prisma.profile.update({
-            data: {
-                pin: pin
-            },
-            where: {
-                id: users.id
-            }
-        });
-
-        return res.status(201).json({
-            status: 201,
-            message: "Created",
-            data: usersPin.pin
         });
     } catch (error) {
         next(error);
@@ -170,9 +90,9 @@ const notification = async (req, res, next) => {
         delete users.password;
 
         return res.status(200).json({
-            status: 200,
+            status: false,
             message: "OK",
-            data: [...notifications]
+            data: notifications
         });
     } catch (error) {
         next(error);
@@ -182,7 +102,5 @@ const notification = async (req, res, next) => {
 module.exports = {
     profile,
     updateProfile,
-    notification,
-    forgotPin,
-    pinValidation
+    notification
 };
